Moved Fine Handlers to another file for better organization
