+++ ccsds/FileAssembler.go
@@ -1,19 +1,16 @@
 package ccsds
 
 import (
-	"archive/zip"
 	"fmt"
 	"github.com/OpenSatelliteProject/SatHelperApp/Logger"
 	"github.com/OpenSatelliteProject/SatHelperApp/XRIT"
 	"github.com/OpenSatelliteProject/SatHelperApp/XRIT/PacketData"
 	"github.com/OpenSatelliteProject/goaec/szwrap"
 	"github.com/mewkiz/pkg/osutil"
-	"io"
 	"io/ioutil"
 	"os"
 	"path"
 	"strconv"
-	"strings"
 )
 
 type FileAssembler struct {
@@ -182,90 +179,5 @@ func (fa *FileAssembler) handleFile(vcid int, filename string) {
 		SLog.Error("Error moving file %s to %s: %s", filename, newPath, err)
 	}
 
-	go fa.postHandleFile(newPath, outBase)
-}
-
-func stripFileHeader(filename string, offset int64) {
-	data, err := ioutil.ReadFile(filename)
-
-	if err != nil {
-		SLog.Error("Error reading file %s: %s", filename, err)
-		return
-	}
-
-	if len(data) < int(offset) {
-		SLog.Error("File %s smaller than offset.", filename)
-		return
-	}
-
-	data = data[offset:]
-
-	err = ioutil.WriteFile(filename, data, os.ModePerm)
-
-	if err != nil {
-		SLog.Error("Error writing file %s: %s", filename, err)
-	}
-}
-
-func (fa *FileAssembler) postHandleFile(filename, outBase string) {
-	xh, err := XRIT.ParseFile(filename)
-
-	if err != nil {
-		SLog.Error("Error parsing file %s: %s", filename, err)
-		_ = os.Remove(filename)
-		return
-	}
-
-	if xh.Compression() == PacketData.ZIP { // Decompress it
-		SLog.Debug("File %s is a zip. Decompressing it.", filename)
-		stripFileHeader(filename, int64(xh.PrimaryHeader.HeaderLength))
-		fa.handleZipFile(filename, outBase)
-	} else if xh.Compression() == PacketData.GIF || xh.Compression() == PacketData.JPEG {
-		SLog.Debug("File %s is a %s.", filename, PacketData.GetCompressionTypeString(xh.Compression()))
-		newName := strings.Replace(filename, ".lrit", PacketData.GetCompressionTypeExtension(xh.Compression()), 1)
-		stripFileHeader(filename, int64(xh.PrimaryHeader.HeaderLength))
-		err = os.Rename(filename, newName)
-		if err != nil {
-			SLog.Error("Error moving file %s to %s: %s", filename, newName, err)
-		}
-	}
-}
-
-func (fa *FileAssembler) handleZipFile(filename, outBase string) {
-	r, err := zip.OpenReader(filename)
-	if err != nil {
-		SLog.Error("Error opening %s as zip file: %s", filename, err)
-		return
-	}
-
-	for _, zipFile := range r.File {
-		SLog.Info("New file decompressed: %s", zipFile.Name)
-		outFile := path.Join(outBase, zipFile.Name)
-		f, err := os.Create(outFile)
-		if err != nil {
-			SLog.Error("Error creating %s: %s", outFile, err)
-			continue
-		}
-
-		r, err := zipFile.Open()
-
-		if err != nil {
-			SLog.Error("Internal ZIP Error: %s", err)
-			_ = f.Close()
-			continue
-		}
-
-		_, err = io.Copy(f, r)
-
-		if err != nil {
-			SLog.Error("Error writing data: %s", err)
-		}
-
-		_ = f.Close()
-		_ = r.Close()
-	}
-
-	_ = r.Close()
-	SLog.Debug("Removing file %s", filename)
-	_ = os.Remove(filename)
+	go PostHandleFile(newPath, outBase)
 }
